+++ src/Bootstrap/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added support for popup messages (toasts).
 * Added the option to disable fog on per visual basis.
 * Fog no longer affects selected residue, ambience, and dynamic density data.
+* Ability to better control layout regions.
 
 # 1.2.1
 * Transform builder no longer requires verbose type annotations that were sometimes required.