+++ src/Plugin/CHANGELOG.md
@@ -1,5 +1,6 @@
 ﻿# 1.2.6
 * Added support for popup messages (toasts).
+* Ability to better control layout regions.
 
 # 1.2.5
 * Updated default pluing spec with Bootstrap.Behaviour.UnselectElementOnRepeatedClick.